Camera should be, at least, moving

+++ src/trait/Updatable.hpp
@@ -8,6 +8,8 @@ namespace trait {
          * Updates the implementer according to how much time elapsed.
          * @param delta Time elapsed between last cycle.
          */
-        virtual void Update(float delta) {};
+        virtual void Update(double delta) {
+
+        };
     };
 }
